Move more early buffered lints to dyn lint diagnostics (6/N)
