+++ compiler/rustc_resolve/messages.ftl
@@ -342,6 +342,10 @@ resolve_note_and_refers_to_the_item_defined_here =
         }
     }
 
+resolve_out_of_scope_macro_calls = cannot find macro `{$path}` in the current scope when looking from {$location}
+    .label = not found from {$location}
+    .help = import `macro_rules` with `use` to make it callable above its definition
+
 resolve_outer_ident_is_not_publicly_reexported =
     {$outer_ident_descr} `{$outer_ident}` is not publicly re-exported
 
@@ -362,11 +366,18 @@ resolve_param_in_ty_of_const_param =
 
 resolve_pattern_doesnt_bind_name = pattern doesn't bind `{$name}`
 
+resolve_private_extern_crate_reexport = extern crate `{$ident}` is private and cannot be re-exported
+    .suggestion = consider making the `extern crate` item publicly accessible
+
 resolve_proc_macro_derive_resolution_fallback = cannot find {$ns_descr} `{$ident}` in this scope
     .label = names from parent modules are not accessible without an explicit import
 
 resolve_proc_macro_same_crate = can't use a procedural macro from the same crate that defines it
     .help = you can define integration tests in a directory named `tests`
+
+resolve_redundant_import_visibility = glob import doesn't reexport anything with visibility `{$import_vis}` because no imported item is public enough
+    .note = the most public imported item is `{$max_vis}`
+    .help = reduce the glob import's visibility or increase visibility of imported items
 
 resolve_reexport_of_crate_public =
     re-export of crate public `{$ident}`
@@ -472,6 +483,9 @@ resolve_unexpected_res_change_ty_to_const_param_sugg =
 
 resolve_unexpected_res_use_at_op_in_slice_pat_with_range_sugg =
     if you meant to collect the rest of the slice in `{$ident}`, use the at operator
+
+resolve_unknown_diagnostic_attribute = unknown diagnostic attribute
+resolve_unknown_diagnostic_attribute_typo_sugg = an attribute with a similar name exists
 
 resolve_unnamed_crate_root_import =
     crate root imports need to be explicitly named: `use crate as name;`
